Verify flac library exists and create mp3 library if needed
The script takes two arguments, the upstream flac library and the
downstream mp3 library. It will ensure that the flac library exists, or
raise an exception.

If the mp3 library doesn't exist, it will create the full path. If the
creation fails for some reason, an exception will be raised.