Resize artwork if present in upstream library
Artwork is expected to be called "album.jpg" and be present in the
second hierarchy level. The downstream library is expected to be mobile,
so the artwork will be resized to 300x300 pixels.

This step uses ImageMagick's convert command.

The name of the artwork file and the reduced size could conceivably be
more flexible in future implementations.