Use a separate class for artwork-related methods
The artfile class should be able to handle resize-move operations on
artwork. The artwork is required to have one of the three allowed
extensions, JPG, JPEG, or PNG. The contents of the file itself are never
checked. If asked to resize, the artwork will be resized using
ImageMagick's convert command.