Perform transcoding in two steps
In the first step, read the MP3 library (if it exists) and determine the
files that need to be transcoded (or resized in case of the artwork
files). Then, only transcode these files.

This commit has also removed the restriction that the library directory
structure had to be in the form artist/album/song. The path is
faithfully recreated in the MP3 library.