Use track number to encode information about disk number
The lame encoder does not know how to encode disk number tags. Further,
some mp3 players don't seem to respect this tag. Now, the track number
will be used to store the information about disk and track numbers.

This commit assumes that albums will never have more than 99 songs,
which is okay as long as we aren't dealing with massive compilation
albums or audio books.

The disk number will henceforth be multiplied by 100 and added to the
track number (if present).