Fix bug in handling path of resized artwork
Unlike the file paths for the FLAC and mp3 file, the artwork paths are
absolute. Simply replacing the FLACDIR with MP3DIR causes trouble when
there is no artwork present. Since there is no existing artwork, but the
artFile path is non-empty, convert probably tries to convert the entire
FLAC directory, leading to an insufficient memory error. Fixed now.