Initialize hash with empty array
Note that

Hash.new Array.new

isn't the same as doing

Hash.new do | hash, key |
    hash[ key ] = Array.new
end

The first uses a new array to initialize lookups for keys that don't
exist in the hash. However, this new array is mutable and doing
something like,

myHash[ newKey ].push element

will actually add element to the initializing array. This is not the
behavior we want.

The second will create a new array anytime an unrecognized key is
presented to the hash. This is the behavior we want.